Split the metadata_template file in multiple files for easy management
